requirements.txt
Django==5.0.2
numpy
# tflite-runtime==2.14.0  <-- Hard to install on Windows, using tensorflow instead
tensorflow
keras-image-helper==0.0.1
Pillow>=9.0
requests>=2.28

# --- Deployment Dependencies (Render) ---
gunicorn>=20.1.0
whitenoise>=6.5.0
python-dotenv>=1.0.0
